Revert "Works with x64 & release"
This reverts commit b48b9473cee1cf391aed166d7d6665c19df2a1f9.

+++ RemoteRendering/RemoteEncoder.cpp
@@ -361,7 +361,7 @@ void RemoteEncoder::createCuda()
 	//Setup Frameprops & Memory...
 	unsigned int widthInBytes = m_EncoderParams->iInputSize[0];
 	unsigned int height = (unsigned int)(m_EncoderParams->iInputSize[1] * 12) >> 3;
-	size_t cuPitch;
+	unsigned int cuPitch;
 	cuRes = cuMemAllocPitch(&dptr, &cuPitch, widthInBytes, height, 16);
 	handleCudaError(cuRes, "MemAllocPitch:");
